move config_valid to util
